refactor: prep for additional doc pages
